Manual merge of https://github.com/SED-ML/sed-ml/pull/145
Some of the changes were already made; a couple of them were missed (Surface log* attributes and a typo).

The biggest discrepancy between them was that there was a sentence added in the 'legend' section that IMO duplicated information about what to call the curves, so I left that out.