Add 'color conventions' section.
We use the SBML color conventions of red==change, blue==links, and blue-in-UML==change, but didn't explain it.  Copied and simplified the version from SBML.

Also added paragraph break in abstract so that we didn't get the line break in the sedml URL.

This commit is the *actual* final JIB submission.  I hope.